backend/requirements.txt
Django==4.2.20
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
stripe==12.1.0
dj-database-url==3.0.1
whitenoise==6.11.0
